+++ core/src/inspect/schema.rs
@@ -1,4 +1,4 @@
-#[derive(Clone, Debug, elephantry_derive::Entity)]
+#[derive(Clone, Debug, Eq, PartialEq, elephantry_derive::Entity)]
 #[elephantry(internal)]
 pub struct Schema {
     pub name: String,