+++ CHANGELOG.md
@@ -18,9 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New `inspect::Relation::kind` field;
 - New `inspect::Relation::persistence` field;
 - Inspect extensions;
+- New `inspect::Domain::constraints` field;
 
 ### Changed
 - `inspect::Relation::ty` is deprecated;
+- `inspect::Domain::constraint` is deprecated;
 
 ### Fixed
 - `u8` conversion;